+++ Changelog.md
@@ -13,6 +13,7 @@
 - Foundry 14: Don't let core Foundry expire turnEnd events.
 - Reinstate `Apply on Outcome` field to allow specific outcomes to affect the test results (e.g having 'test.damage' or 'test.weaponAP' in the changes).
 - Transferring an effect to an Actor should now generate a scrolling status on the token (_displayScrollingStatus).
+- `Active Defense` and the two defenses in the Macro Hub now use `system.defenses.all.mod` (or `system.defenses.physical.mod`)
 
 ### Other Improvements
 - When `Concentration` is lost from an Actor, any effects linked to that concentration (on any Actors) are immediately removed.